refactor(desktop): split mastra user message pieces
