+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/ChatMastraPane/ChatMastraInterface/components/ChatMastraMessageList/components/UserMessage/components/UserMessageText/index.ts
@@ -1,0 +1,1 @@
+export { UserMessageText } from "./UserMessageText";
