+++ apps/desktop/src/renderer/screens/main/components/WorkspaceView/ContentView/TabsContent/TabView/ChatMastraPane/ChatMastraInterface/components/ChatMastraMessageList/components/UserMessage/types.ts
@@ -1,0 +1,7 @@
+import type { UseMastraChatDisplayReturn } from "@superset/chat-mastra/client";
+
+export type MastraMessage = NonNullable<
+	UseMastraChatDisplayReturn["messages"]
+>[number];
+
+export type MastraMessagePart = MastraMessage["content"][number];
